# system-design-lab-api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ChoiceResponse:
    properties:
      consequences:
        items:
          $ref: '#/definitions/dto.ConsequencesResponse'
        type: array
      explanations:
        additionalProperties:
          type: string
        type: object
      id:
        type: string
      impacts:
        items:
          $ref: '#/definitions/dto.ImpactResponse'
        type: array
      isCorrect:
        type: boolean
      label:
        type: string
      nextStepId:
        type: string
    type: object
...
  dto.ConsequencesResponse:
    properties:
      key:
        type: string
      value:
        type: boolean
    type: object
...
  dto.CreateStepRequest:
    properties:
      context:
        maxLength: 500
        type: string
      orderIndex:
        type: integer
      question:
        minLength: 3
        type: string
      scenarioID:
        type: string
    required:
    - question
    type: object
...
  dto.ImpactResponse:
    properties:
      metric:
        type: string
      type:
        type: string
      value:
        type: number
    type: object
...
  dto.SessionResponse:
    properties:
      currentStep:
        $ref: '#/definitions/dto.StepResponse'
      flags:
        additionalProperties:
          type: boolean
        type: object
      id:
        type: string
      metrics:
        additionalProperties:
          format: float64
          type: number
        type: object
      scenarioId:
        type: string
      status:
        type: string
    type: object
  dto.SessionSummaryResponse:
    properties:
      choices:
        items:
          $ref: '#/definitions/dto.UserChoiceSummary'
        type: array
      createdAt:
        type: string
      flags:
        additionalProperties:
          type: boolean
        type: object
      id:
        type: string
      metrics:
        additionalProperties:
          format: float64
          type: number
        type: object
      scenarioId:
        type: string
      status:
        type: string
    type: object
  dto.StartSessionRequest:
    properties:
      scenarioId:
        type: string
      userId:
        type: string
    required:
    - scenarioId
    - userId
    type: object
  dto.StepPaginationResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.StepResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  dto.StepResponse:
    properties:
      choices:
        items:
          $ref: '#/definitions/dto.ChoiceResponse'
        type: array
      context:
        type: string
      id:
        type: string
      question:
        type: string
    type: object
  dto.SubmitChoiceRequest:
    properties:
      choiceId:
        type: string
    required:
    - choiceId
    type: object
  dto.SubmitChoiceResponse:
    properties:
      explanations:
        additionalProperties:
          type: string
        type: object
      feedback:
        type: string
      isCompleted:
        type: boolean
      isCorrect:
        type: boolean
      metrics:
        additionalProperties:
          format: float64
          type: number
        type: object
      nextStep:
        $ref: '#/definitions/dto.StepResponse'
    type: object
...
  dto.UserChoiceSummary:
    properties:
      choiceId:
        type: string
      explanations:
        additionalProperties:
          type: string
        type: object
      isCorrect:
        type: boolean
      label:
        type: string
      question:
        type: string
      stepId:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Guided System Design Playground — learn system design through decisions
    and consequences.
  title: System Design Lab API
  version: "1.0"
paths:
...
  /sessions:
    post:
      consumes:
      - application/json
      parameters:
      - description: Session payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.StartSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Start a new session
      tags:
      - sessions
  /sessions/{id}:
    get:
      parameters:
      - description: Session UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SessionResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get session by ID
      tags:
      - sessions
  /sessions/{id}/submit:
    post:
      consumes:
      - application/json
      parameters:
      - description: Session UUID
        in: path
        name: id
        required: true
        type: string
      - description: Choice payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SubmitChoiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubmitChoiceResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Submit a choice for a session
      tags:
      - sessions
  /sessions/{id}/summary:
    get:
      parameters:
      - description: Session UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SessionSummaryResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get session summary
      tags:
      - sessions
  /steps:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StepPaginationResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List steps
      tags:
      - steps
    post:
      consumes:
      - application/json
      parameters:
      - description: Step payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateStepRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.StepResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a step
      tags:
      - steps
swagger: "2.0"